new changes

+++ database/migrations/2025_12_13_085212_car_type_surcharge_times.php
@@ -27,7 +27,7 @@ return new class extends Migration
             $table->time('end_time');
 
             // surcharge value
-            $table->decimal('amount', 8, 2);
+            $table->unsignedInteger('amount');
 
             $table->timestamps();
         });
